Agent_csv/requirements.txt
langchain-core>=0.1.0
langgraph>=0.1.0
pandas>=1.5.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0 